+++ R/ald_demo.R
@@ -16,8 +16,6 @@
 #' * `country_of_domicile` (character): Country where company is registered.
 #' * `emission_factor` (double): Company level emission factor of the
 #' technology.
-#' * `is_ultimate_listed_owner` (logical): Flag if company is the listed
-#' ultimate parent.
 #' * `name_company` (character): The name of the company owning the asset.
 #' * `number_of_assets` (integer): Number of assets of a given technology owned
 #' by the company.