+++ NEWS.md
@@ -1,5 +1,4 @@
 # r2dii.data 0.0.3.9000
 
-* The column `is_ultimate_owner` was removed from `ald_demo` (#22). That column was unused and unavailable to banks.
+* Remove columns `is_ultimate_owner` and `is_ultimate_listed_owner` from `ald_demo` (#22). They are unnecessary and unavailable for banks.
 
-* Added a `NEWS.md` file to track changes to the package.